+++ src/resources/lib/modules/system.py
@@ -44,9 +44,6 @@ class system:
 
     ENABLED = False
     KERNEL_CMD = None
-    UPDATE_REQUEST_URL = None
-    UPDATE_DOWNLOAD_URL = None
-    LOCAL_UPDATE_DIR = None
     XBMC_RESET_FILE = None
     LIBREELEC_RESET_FILE = None
     KEYBOARD_INFO = None
@@ -132,111 +129,6 @@ class system:
                             },
                         },
                     },
-                'update': {
-                    'order': 3,
-                    'name': 32013,
-                    'settings': {
-                        'AutoUpdate': {
-                            'name': 32014,
-                            'value': 'auto',
-                            'action': 'set_auto_update',
-                            'type': 'multivalue',
-                            'values': ['auto', 'manual'],
-                            'InfoText': 714,
-                            'order': 1,
-                            },
-                        'SubmitStats': {
-                            'name': 32021,
-                            'value': '1',
-                            'action': 'set_value',
-                            'type': 'bool',
-                            'InfoText': 772,
-                            'order': 2,
-                            },
-                        'UpdateNotify': {
-                            'name': 32365,
-                            'value': '1',
-                            'action': 'set_value',
-                            'type': 'bool',
-                            'InfoText': 715,
-                            'order': 3,
-                            },
-                        'ShowCustomChannels': {
-                            'name': 32016,
-                            'value': '0',
-                            'action': 'set_custom_channel',
-                            'type': 'bool',
-                            'parent': {
-                                'entry': 'AutoUpdate',
-                                'value': ['manual'],
-                                },
-                            'InfoText': 761,
-                            'order': 4,
-                            },
-                        'CustomChannel1': {
-                            'name': 32017,
-                            'value': '',
-                            'action': 'set_custom_channel',
-                            'type': 'text',
-                            'parent': {
-                                'entry': 'ShowCustomChannels',
-                                'value': ['1'],
-                                },
-                            'InfoText': 762,
-                            'order': 5,
-                            },
-                        'CustomChannel2': {
-                            'name': 32018,
-                            'value': '',
-                            'action': 'set_custom_channel',
-                            'type': 'text',
-                            'parent': {
-                                'entry': 'ShowCustomChannels',
-                                'value': ['1'],
-                                },
-                            'InfoText': 762,
-                            'order': 6,
-                            },
-                        'CustomChannel3': {
-                            'name': 32019,
-                            'value': '',
-                            'action': 'set_custom_channel',
-                            'type': 'text',
-                            'parent': {
-                                'entry': 'ShowCustomChannels',
-                                'value': ['1'],
-                                },
-                            'InfoText': 762,
-                            'order': 7,
-                            },
-                        'Channel': {
-                            'name': 32015,
-                            'value': '',
-                            'action': 'set_channel',
-                            'type': 'multivalue',
-                            'parent': {
-                                'entry': 'AutoUpdate',
-                                'value': ['manual'],
-                                },
-                            'values': [],
-                            'InfoText': 760,
-                            'order': 8,
-                            },
-                        'Build': {
-                            'name': 32020,
-                            'value': '',
-                            'action': 'do_manual_update',
-                            'type': 'multivalue',
-                            'parent': {
-                                'entry': 'AutoUpdate',
-                                'value': ['manual'],
-                                },
-                            'values': [],
-                            'InfoText': 770,
-                            'order': 9,
-                            },
-                        },
-                    },
                 'backup': {
                     'order': 7,
                     'name': 32371,
@@ -307,7 +199,6 @@ class system:
 
             self.keyboard_layouts = False
             self.nox_keyboard_layouts = False
-            self.last_update_check = 0
             self.arrVariants = {}
             self.oe.dbg_log('system::__init__', 'exit_function', 0)
         except Exception, e:
@@ -321,7 +212,6 @@ class system:
             self.set_hostname()
             self.set_keyboard_layout()
             self.set_hw_clock()
-            self.set_auto_update()
             del self.is_service
             self.oe.dbg_log('system::start_service', 'exit_function', 0)
         except Exception, e:
@@ -459,53 +349,6 @@ class system:
                 self.struct['ident']['settings']['hostname']['value'] = value
             else:
                 self.struct['ident']['settings']['hostname']['value'] = self.oe.DISTRIBUTION
-
-            # AutoUpdate
-
-            value = self.oe.read_setting('system', 'AutoUpdate')
-            if not value is None:
-                self.struct['update']['settings']['AutoUpdate']['value'] = value
-            value = self.oe.read_setting('system', 'SubmitStats')
-            if not value is None:
-                self.struct['update']['settings']['SubmitStats']['value'] = value
-            value = self.oe.read_setting('system', 'UpdateNotify')
-            if not value is None:
-                self.struct['update']['settings']['UpdateNotify']['value'] = value
-            if os.path.isfile('%s/SYSTEM' % self.LOCAL_UPDATE_DIR):
-                self.update_in_progress = True
-
-            # Manual Update
-
-            value = self.oe.read_setting('system', 'Channel')
-            if not value is None:
-                self.struct['update']['settings']['Channel']['value'] = value
-            value = self.oe.read_setting('system', 'ShowCustomChannels')
-            if not value is None:
-                self.struct['update']['settings']['ShowCustomChannels']['value'] = value
-
-            value = self.oe.read_setting('system', 'CustomChannel1')
-            if not value is None:
-                self.struct['update']['settings']['CustomChannel1']['value'] = value
-            value = self.oe.read_setting('system', 'CustomChannel2')
-            if not value is None:
-                self.struct['update']['settings']['CustomChannel2']['value'] = value
-            value = self.oe.read_setting('system', 'CustomChannel3')
-            if not value is None:
-                self.struct['update']['settings']['CustomChannel3']['value'] = value
-
-            self.update_json = self.build_json()
-
-            self.struct['update']['settings']['Channel']['values'] = self.get_channels()
-            self.struct['update']['settings']['Build']['values'] = self.get_available_builds()
-
-            # AutoUpdate = manual by environment var.
-
-            if os.path.exists('/dev/.update_disabled'):
-                self.update_disabled = True
-                self.struct['update']['hidden'] = 'true'
-                self.struct['update']['settings']['AutoUpdate']['value'] = 'manual'
-                self.struct['update']['settings']['UpdateNotify']['value'] = '0'
-            self.oe.dbg_log('system::load_values', 'exit_function', 0)
         except Exception, e:
             self.oe.dbg_log('system::load_values', 'ERROR: (' + repr(e) + ')')
 
@@ -607,22 +450,6 @@ class system:
             self.oe.set_busy(0)
             self.oe.dbg_log('system::set_hostname', 'ERROR: (' + repr(e) + ')')
 
-    def set_auto_update(self, listItem=None):
-        try:
-            self.oe.dbg_log('system::set_auto_update', 'enter_function', 0)
-            if not listItem == None:
-                self.set_value(listItem)
-            if not hasattr(self, 'update_disabled'):
-                if not hasattr(self, 'update_thread'):
-                    self.update_thread = updateThread(self.oe)
-                    self.update_thread.start()
-                else:
-                    self.update_thread.wait_evt.set()
-                self.oe.dbg_log('system::set_auto_update', unicode(self.struct['update']['settings']['AutoUpdate']['value']), 1)
-            self.oe.dbg_log('system::set_auto_update', 'exit_function', 0)
-        except Exception, e:
-            self.oe.dbg_log('system::set_auto_update', 'ERROR: (' + repr(e) + ')')
-
     def get_keyboard_layouts(self):
         try:
             self.oe.dbg_log('system::get_keyboard_layouts', 'enter_function', 0)
@@ -690,205 +517,6 @@ class system:
         except Exception, e:
             self.oe.dbg_log('system::get_keyboard_layouts', 'ERROR: (' + repr(e) + ')')
 
-    def set_channel(self, listItem=None):
-        try:
-            self.oe.dbg_log('system::set_channel', 'enter_function', 0)
-            if not listItem == None:
-                self.set_value(listItem)
-            self.struct['update']['settings']['Build']['values'] = self.get_available_builds()
-            self.oe.dbg_log('system::set_channel', 'exit_function', 0)
-        except Exception, e:
-            self.oe.dbg_log('system::set_channel', 'ERROR: (' + repr(e) + ')')
-
-    def set_custom_channel(self, listItem=None):
-        try:
-            self.oe.dbg_log('system::set_custom_channel', 'enter_function', 0)
-            if not listItem == None:
-                self.set_value(listItem)
-            if listItem.getProperty('entry') != 'ShowCustomChannels':
-                if self.get_json(listItem.getProperty('value')) is None:
-                    xbmcDialog = xbmcgui.Dialog()
-                    xbmcDialog.ok('LibreELEC Update', self.oe._(32191).encode('utf-8'))
-                    xbmcDialog = None
-                    del xbmcDialog
-            self.update_json = self.build_json()
-            self.struct['update']['settings']['Channel']['values'] = self.get_channels()
-            if not self.struct['update']['settings']['Channel']['values'] is None:
-                if not self.struct['update']['settings']['Channel']['value'] in self.struct['update']['settings']['Channel']['values']:
-                    self.struct['update']['settings']['Channel']['value'] = None
-            self.struct['update']['settings']['Build']['values'] = self.get_available_builds()
-            self.oe.dbg_log('system::set_custom_channel', 'exit_function', 0)
-        except Exception, e:
-            self.oe.dbg_log('system::set_custom_channel', 'ERROR: (' + repr(e) + ')')
-
-    def get_channels(self):
-        try:
-            self.oe.dbg_log('system::get_channels', 'enter_function', 0)
-            channels = []
-            self.oe.dbg_log('system::get_channels', unicode(self.update_json), 0)
-            if not self.update_json is None:
-                for channel in self.update_json:
-                    channels.append(channel)
-            self.oe.dbg_log('system::get_channels', 'exit_function', 0)
-            return channels
-        except Exception, e:
-            self.oe.dbg_log('system::get_channels', 'ERROR: (' + repr(e) + ')')
-
-    def do_manual_update(self, listItem=None):
-        try:
-            self.oe.dbg_log('system::do_manual_update', 'enter_function', 0)
-            channel = self.struct['update']['settings']['Channel']['value']
-            regex = re.compile(self.update_json[channel]['prettyname_regex'])
-            longname = '-'.join([self.oe.DISTRIBUTION, self.oe.ARCHITECTURE, self.oe.VERSION])
-            if regex.search(longname):
-                version = regex.findall(longname)[0]
-            else:
-                version = self.oe.VERSION
-            if not listItem == None:
-                self.struct['update']['settings']['Build']['value'] = listItem.getProperty('value')
-            if self.struct['update']['settings']['Build']['value'] != '':
-                self.update_file = self.update_json[self.struct['update']['settings']['Channel']['value']]['url'] + self.get_available_builds(self.struct['update']['settings']['Build']['value'])
-                xbmcDialog = xbmcgui.Dialog()
-                answer = xbmcDialog.yesno('LibreELEC Update', self.oe._(32188).encode('utf-8') + ':  ' + version.encode('utf-8'),
-                                      self.oe._(32187).encode('utf-8') + ':  ' + self.struct['update']['settings']['Build']['value'].encode('utf-8'),
-                                      self.oe._(32180).encode('utf-8'))
-                xbmcDialog = None
-                del xbmcDialog
-                if answer:
-                    self.update_in_progress = True
-                    self.do_autoupdate()
-            self.struct['update']['settings']['Build']['value'] = ''
-            self.oe.dbg_log('system::do_manual_update', 'exit_function', 0)
-        except Exception, e:
-            self.oe.dbg_log('system::do_manual_update', 'ERROR: (' + repr(e) + ')')
-
-    def get_json(self, url=None):
-        try:
-            self.oe.dbg_log('system::get_json', 'enter_function', 0)
-            if url is None:
-                url = self.UPDATE_DOWNLOAD_URL % ('releases', 'releases.json')
-            if url.split('/')[-1] != 'releases.json':
-                url = url + '/releases.json'
-            data = self.oe.load_url(url)
-            if not data is None:
-                update_json = json.loads(data)
-            else:
-                update_json = None
-            self.oe.dbg_log('system::get_json', 'exit_function', 0)
-            return update_json
-        except Exception, e:
-            self.oe.dbg_log('system::get_json', 'ERROR: (' + repr(e) + ')')
-
-    def build_json(self):
-        try:
-            self.oe.dbg_log('system::build_json', 'enter_function', 0)
-            update_json = self.get_json()
-            if self.struct['update']['settings']['ShowCustomChannels']['value'] == '1':
-                custom_urls = []
-                for i in 1,2,3:
-                    custom_urls.append(self.struct['update']['settings']['CustomChannel' + str(i)]['value'])
-                for custom_url in custom_urls:
-                    if custom_url != '':
-                        custom_update_json = self.get_json(custom_url)
-                        if not custom_update_json is None:
-                            for channel in custom_update_json:
-                                update_json[channel] = custom_update_json[channel]
-            self.oe.dbg_log('system::build_json', 'exit_function', 0)
-            return update_json
-        except Exception, e:
-            self.oe.dbg_log('system::build_json', 'ERROR: (' + repr(e) + ')')
-
-    def get_available_builds(self, shortname=None):
-        try:
-            self.oe.dbg_log('system::get_available_builds', 'enter_function', 0)
-            channel = self.struct['update']['settings']['Channel']['value']
-            update_files = []
-            build = None
-            if not self.update_json is None:
-                if channel != '':
-                    if channel in self.update_json:
-                        regex = re.compile(self.update_json[channel]['prettyname_regex'])
-                        if self.oe.ARCHITECTURE in self.update_json[channel]['project']:
-                            for i in sorted(self.update_json[channel]['project'][self.oe.ARCHITECTURE]['releases'], key=int, reverse=True):
-                                if shortname is None:
-                                    update_files.append(regex.findall(self.update_json[channel]['project'][self.oe.ARCHITECTURE]['releases'][i]['file']['name'])[0].strip('.tar'))
-                                else:
-                                    build = self.update_json[channel]['project'][self.oe.ARCHITECTURE]['releases'][i]['file']['name']
-                                    if shortname in build:
-                                        break
-            self.oe.dbg_log('system::get_available_builds', 'exit_function', 0)
-            if build is None:
-                return update_files
-            else:
-                return build
-        except Exception, e:
-            self.oe.dbg_log('system::get_available_builds', 'ERROR: (' + repr(e) + ')')
-
-    def check_updates_v2(self, force=False):
-        try:
-            self.oe.dbg_log('system::check_updates_v2', 'enter_function', 0)
-            if hasattr(self, 'update_in_progress'):
-                self.oe.dbg_log('system::check_updates_v2', 'Update in progress (exit)', 0)
-                return
-            if self.struct['update']['settings']['SubmitStats']['value'] == '1':
-                systemid = self.oe.SYSTEMID
-            else:
-                systemid = "NOSTATS"
-            if self.oe.BUILDER_VERSION:
-                version = self.oe.BUILDER_VERSION
-            else:
-                version = self.oe.VERSION
-            url = '%s?i=%s&d=%s&pa=%s&v=%s&f=%s' % (
-                self.UPDATE_REQUEST_URL,
-                self.oe.url_quote(systemid),
-                self.oe.url_quote(self.oe.DISTRIBUTION),
-                self.oe.url_quote(self.oe.ARCHITECTURE),
-                self.oe.url_quote(version),
-                self.oe.url_quote(self.hardware_flags),
-                )
-            if self.oe.BUILDER_NAME:
-               url += '&b=%s' % self.oe.url_quote(self.oe.BUILDER_NAME)
-
-            self.oe.dbg_log('system::check_updates_v2', 'URL: %s' % url, 0)
-            update_json = self.oe.load_url(url)
-            self.oe.dbg_log('system::check_updates_v2', 'RESULT: %s' % repr(update_json), 0)
-            if update_json != '':
-                update_json = json.loads(update_json)
-                self.last_update_check = time.time()
-                if 'update' in update_json['data'] and 'folder' in update_json['data']:
-                    self.update_file = self.UPDATE_DOWNLOAD_URL % (update_json['data']['folder'], update_json['data']['update'])
-                    if self.struct['update']['settings']['UpdateNotify']['value'] == '1':
-                        self.oe.notify(self.oe._(32363).encode('utf-8'), self.oe._(32364).encode('utf-8'))
-                    if self.struct['update']['settings']['AutoUpdate']['value'] == 'auto' and force == False:
-                        self.update_in_progress = True
-                        self.do_autoupdate(None, True)
-            self.oe.dbg_log('system::check_updates_v2', 'exit_function', 0)
-        except Exception, e:
-            self.oe.dbg_log('system::check_updates_v2', 'ERROR: (' + repr(e) + ')')
-
-    def do_autoupdate(self, listItem=None, silent=False):
-        try:
-            self.oe.dbg_log('system::do_autoupdate', 'enter_function', 0)
-            if hasattr(self, 'update_file'):
-                if not os.path.exists(self.LOCAL_UPDATE_DIR):
-                    os.makedirs(self.LOCAL_UPDATE_DIR)
-                downloaded = self.oe.download_file(self.update_file, self.oe.TEMP + 'update_file', silent)
-                if not downloaded is None:
-                    self.update_file = self.update_file.split('/')[-1]
-                    if self.struct['update']['settings']['UpdateNotify']['value'] == '1':
-                        self.oe.notify(self.oe._(32363), self.oe._(32366))
-                    shutil.move(self.oe.TEMP + 'update_file', self.LOCAL_UPDATE_DIR + self.update_file)
-                    subprocess.call('sync', shell=True, stdin=None, stdout=None, stderr=None)
-                    if silent == False:
-                        self.oe.winOeMain.close()
-                        time.sleep(1)
-                        xbmc.executebuiltin('Reboot')
-                else:
-                    delattr(self, 'update_in_progress')
-
-            self.oe.dbg_log('system::do_autoupdate', 'exit_function', 0)
-        except Exception, e:
-            self.oe.dbg_log('system::do_autoupdate', 'ERROR: (' + repr(e) + ')')
 
     def set_hw_clock(self):
         try:
@@ -962,15 +590,15 @@ class system:
                 self.oe.set_busy(0)
 
             xbmcDialog = xbmcgui.Dialog()
-            bckDir = xbmcDialog.browse( 0, 
+            bckDir = xbmcDialog.browse( 0,
                                         self.oe._(32371).encode('utf-8'),
-                                        'files', 
-                                        '', 
-                                        False, 
-                                        False, 
+                                        'files',
+                                        '',
+                                        False,
+                                        False,
                                         self.BACKUP_DESTINATION )
 
-            if bckDir and os.path.exists(bckDir): 
+            if bckDir and os.path.exists(bckDir):
                 # free space check
                 try:
                     folder_stat = os.statvfs(bckDir)
@@ -982,7 +610,7 @@ class system:
                         return 0
                 except:
                     pass
-                    
+
                 self.backup_dlg = xbmcgui.DialogProgress()
                 self.backup_dlg.create('LibreELEC', self.oe._(32375).encode('utf-8'), ' ', ' ')
                 if not os.path.exists(self.BACKUP_DESTINATION):
@@ -1005,16 +633,16 @@ class system:
             self.oe.dbg_log('system::do_restore', 'enter_function', 0)
             copy_success = 0
             xbmcDialog = xbmcgui.Dialog()
-            restore_file_path = xbmcDialog.browse( 1, 
+            restore_file_path = xbmcDialog.browse( 1,
                                               self.oe._(32373).encode('utf-8'),
-                                              'files', 
-                                              '??????????????.tar', 
-                                              False, 
-                                              False, 
+                                              'files',
+                                              '??????????????.tar',
+                                              False,
+                                              False,
                                               self.BACKUP_DESTINATION )
-            
+
             restore_file_name = restore_file_path.split('/')[-1]
-            
+
             if not os.path.exists(self.RESTORE_DIR):
                 os.makedirs(self.RESTORE_DIR)
             else:
@@ -1208,46 +836,3 @@ class system:
             self.oe.dbg_log('system::wizard_set_hostname', 'exit_function', 0)
         except Exception, e:
             self.oe.dbg_log('system::wizard_set_hostname', 'ERROR: (' + repr(e) + ')')
-
-
-class updateThread(threading.Thread):
-
-    def __init__(self, oeMain):
-        try:
-            oeMain.dbg_log('system::updateThread::__init__', 'enter_function', 0)
-            self.oe = oeMain
-            self.stopped = False
-            self.wait_evt = threading.Event()
-            threading.Thread.__init__(self)
-            self.oe.dbg_log('system::updateThread', 'Started', 1)
-            self.oe.dbg_log('system::updateThread::__init__', 'exit_function', 0)
-        except Exception, e:
-            self.oe.dbg_log('system::updateThread::__init__', 'ERROR: (' + repr(e) + ')')
-
-    def stop(self):
-        try:
-            self.oe.dbg_log('system::updateThread::stop()', 'enter_function', 0)
-            self.stopped = True
-            self.wait_evt.set()
-            self.oe.dbg_log('system::updateThread::stop()', 'exit_function', 0)
-        except Exception, e:
-            self.oe.dbg_log('system::updateThread::stop()', 'ERROR: (' + repr(e) + ')')
-
-    def run(self):
-        try:
-            self.oe.dbg_log('system::updateThread::run', 'enter_function', 0)
-            while self.stopped == False:
-                if not xbmc.Player().isPlaying():
-                    self.oe.dictModules['system'].check_updates_v2()
-                if not hasattr(self.oe.dictModules['system'], 'update_in_progress'):
-                    self.wait_evt.wait(21600)
-                else:
-                    self.oe.notify(self.oe._(32363).encode('utf-8'), self.oe._(32364).encode('utf-8'))
-                    self.wait_evt.wait(3600)
-                self.wait_evt.clear()
-            self.oe.dbg_log('system::updateThread', 'Stopped', 1)
-            self.oe.dbg_log('system::updateThread::run', 'exit_function', 0)
-        except Exception, e:
-            self.oe.dbg_log('system::updateThread::run', 'ERROR: (' + repr(e) + ')')
-
-
